reorganize the writer module
