+++ icelake/src/io/file_writer/mod.rs
@@ -5,5 +5,3 @@ pub mod position_delete_writer;
 pub use position_delete_writer::*;
 pub mod equality_delete_writer;
 pub use equality_delete_writer::*;
-pub mod equality_delta_writer;
-pub use equality_delta_writer::*;
